requirements.txt
google-generativeai==0.3.2
Pillow==10.2.0
pyautogui==0.9.54
python-dotenv==1.0.1
pynput==1.7.6
pyperclip==1.8.2
opencv-python==4.9.0.80
numpy==1.26.4
PyQt5==5.15.10 